Set default workspace in user directory
The default workspace was the current directory, but this isn't very helpful when an executable is created and run from a system protected directory. On a Windows machine it is put in the "Documents" folder, otherwise in the home directory. This commit also creates the workspace directory if it does not exist.