Add six import for vispy to not complain when building an executable
This wasn't needed in the past and I'm not sure what changed, but this seemed to fix the build. After running `pyinstaller sift.spec` then running the `dist/bin/SIFT.exe` file would raise an error from vispy about the version of six. It is mentioned in this vispy github issue:
https://github.com/vispy/vispy/issues/1074