Warn user when xrit cannot be decompressed (WORKAROUND)
To decompress compressed SEVIRI HRIT files, the environment variable
XRIT_DECOMPRESS_PATH needs to point to the xRITDecompress binary. If
this is not the case, loading compressed XRIT data will fail and lead to
a crash of SIFT when finishing the open file dialog and SIFT tries to
load data in background.

To estimate whether reading might fail already during file selection, a
workaround is introduced that checks if a specific reader (now: only
seviri_l1b_frit) is being used and then if certain meta data says that
compression is being used. In case the check is fulfilled, SIFT tries to
access the required environment variable or informs the user about the
issue while preventing loading of these files.

NOTE: this is considered a workaround since the check might not exactly
determine if loading will succeed, and prevent loading in case a failure
is assumed.

SUGGESTION: Satpy might be extended to inform about loading issue as
soon as creating a scene requiring unaccessible external tools, so that
client applications can consider this information.