Enable to import Satpy from configured location
by prepending a custom PathFinder into the system package search path
and instruct it to look for satpy at the location given by the config
setting 'satpy_import_path'.

Also add an according config file which - if used - would set that
setting to 'None', which is interpreted as to don't change the default
import behaviour for Satpy.