Create configuration to control Caching
Two options are provided:
* USE_INVENTORY_DB - whether the inventory database shall be used.
  This will define, whether a CachingWorkspace or a SimpleWorkspace
  (implemented in next commits) shall be used
* CLEANUP_FILE_CACHE - whether intermediate *.image files will be
  kept in the .cache folder. This setting has only an effect with a
  SimpleWorkspace, i.e. when USE_INVENTORY_DB is False.

It is not feasible (nor implemented) to switch between operation with
and without file/DB based caching, Therefore these options cannot be
changed at runtime (i.e. it is "static") but only via configuration for
now (in the future a command line option may be added).

Because of its static nature the configuration setting can be and is
stored in the globally accessible "constant" USE_FILE_CACHING, which is
initialized at startup.