Implement new-data-heartbeat for watchdog
When a new dataset is added as layer it's datetime is written as a
"heartbeat" into a file with configurable location (defaulting to
'watchdog.txt' in CACHE_DIR).

The writing operation is atomic: First MTG-SIFT will create a journal
and after a fsync the file will be renamed to watchdog.txt. This way the
watchdog doesn't read a file still in write progress.

The watchdog emits a warning by default if the dataset is older than 10
minutes. The max age of the dataset and the search interval can be
customized using CLI arguments. Furthermore a verbose mode can be
activated.