Improve log message when tiled image isn't displayed
When the TiledGeolocatedImage is panned to leave the canvas completely a
ValueError exception is raised to notify about this and also a WARNING log
message is generated (it can be discussed, whether this case is worth a
warning, but...). This commit enhances the wording and adds the
ValueError exception message to the log message in the hope to reduce
confusion about its meaning.