Implement auto restart for monitoring application
The watchdog monitors the heartbeat file in the cache directory
of the MTG-SIFT application. This file contains the creation date
of the last imported dataset and the main PID. This PID can then
be used by the watchdog to check whether the MTG-SIFT application
is still alive or send a restart request.

A restart request will be sent if one of the following conditions
is true:
  - imported dataset is too old
  - application didn't import a dataset in a specified time interval
  - uses too much memory
  - was not restarted for a long time

The conversion to int in line 64 and 72 is important:

>>> "0" == 0
False

The 'get' function from donfig returns a str or a dict. Therefore
all integers in YAML will be converted to a str. The above output
from the Python REPL demonstrates, that the true branch will only
be taken if the config key isn't present. In that case the default
value '0' will be used, which has the type int.