Enforce padding and flipping for Satpy loader
Let Satpy provide a full disk image padded with NaN when only parts of
of it are loaded (e.g. only some stripes of an SEVIRI HRIT file set or
an FCI netCDF file set) are loaded. Without this change the loaded
sections would be moved to the south.

This allows MTG-SIFT to just overlay the various datasets without
worrying about positioning.

Futhermore Satpy is instructed to flip all read data to conform to the
standard data layout (upper_right_corner="NE").