Remove "private" methods from Base/SimpleWorkspace
which are abstract (in BaseWorkspace) or have empty ("pass")
implementation (in SimpleWorkspace).

Maintenance:

This commit cleans up unused code which was left in Base/SimpleWorkspace
in the previous commit on purpose to have a better starting point for
enhancing SimpleWorkspace. Revert this commit if useful.

One method, _create_product_from_array(), is left since it is called in
the BaseWorkspace.create_algebraic_composite().

FIXME: for SimpleWorkspace, create_algebraic_composite() is broken
(throwing exception) because
SimpleWorkspace._create_product_from_array() doesn't return required
data.