TMP: Integrate catalogue evaluation at application start
Temporarily adds a configuration option "auto_load" to activate
automatic loading of data according to a catalogue query configuration
when the application is started.

This commit is only for having an intermediate possibility to test the
integrated catalogue functionality. The commit is intended to be
reversed when superseded by the actual auto update mode implemenatation.