Fix issue with RGB layer removal when caching is off
The implementation for removing an RGB layer didn't take into account,
that these do not have own 'content' or 'product', thus its uuid is not
listed in the according dictionaries. This patch avoids the KeyError
exception in SimpleWorkspace which happens when that uuid is popped from
these dictionaries without precautions.