Let the auto update GUI display times in UTC
Display all entries in the current datetimes display (shown When
auto_update_mode is True) showing their timezone.

This reveals a BUG (issue in legacy code or even Satpy):

For displayed point in time when the last update cycle finished loading
new data and the current time display (quasi providing a watch) "now" is
taken for timezone UTC (i.e. with datetime.now(tz=timezone.utc)) - this
is OK and how it should be done!

But: the data time (Info.SCHED_TIME) stored in the (MTG-)SIFT document
does not have a timezone set. This is problematic and needs
investigation.