Make exception for missing catalogue more explicit
When no catalogue query can be found in the configuration, a
RuntimeError exception with proper message is raised.

Previous to this commit either a TypeError or an IndexError were raised
depending on what is missing (no catalogue config at all or catalogue
without any entry in its (query) list), which doesn't help the user to
understand the cause of the issue.