Implement grouping mode functionality
to allow the choice between the grouping modes
1) by group keys
2) keep separate
3) merge all

Currently for data with AreaDefinition only (1) is possible because at
least the SEVIRI Satpy reader does not allow for (2) and there is no use
case known for (3) for that data (nor for FCI FDHSI).

If no group keys are defined - neither in the MTG-SIFT reader's
configuration nor in the Satpy reader's configuration - (1) behaves like
(2), but the existence of any group keys cannot be easily checked
beforehand.

Implementation details:
- by taking advantage of the override for QWizard.initializePage() the
  slot 'on_page_changed' could be removed