Separate usage of LayerSet's next_frame and tick methods
Update of animation slider position as well as checkmark in
layer list tied into LayerSet's next_frame method. This method also
calls itself recursively by being passed to itself as a kwarg. This
leads to some unexpected behavior when expecting a linear progression
in time upon the animation timer timing out and emitting it's relevant
signal. The tick method in the LayerSet class has been refactored to
trigger the time managers tick method and set the required layer
visibilities afterward. It is now connected as a slot to the animation
timers timeout signal, as is the next_frame method.