Persist time transformer policy and time transformer
On Policy instantiation it is not assumed that a sound collection
is present. Instead the policy provides a QT slot allowing it to
react to changes in the collection in an event driven manner.

Thus the Policy's lifetime is no longer coupled to the validity of
the collection, allowing the policy to be instantiated in the
_init_collection method of the TimeManager class, called in TimeManager's __init__.

TimeManager also does not provide a setter definition for collection
anymore, however it does still provide collection as a property to allow
other classes to access collection and signal its change when necessary.

To that end the Collection now derives from QObject, allowing it to emit
QT signals.