NEEDS REVIEW: Fix suspicious SatpyImporter internal
The calculation of the local variable num_contents in
SatpyImporter.begin_import_products() looked suspicious.
The variable was introduced initially in commit
"Add contour visuals to SGM and contour calculation to importer"
(bba8d73f8a37b5c049e39d833c887e7b3783e9ee)
and is used to calculate a completion value (most likely not that
important) but also for control flow.

Especially because of the latter it would make sense to get feedback
from the original author of that line (David Hoese), whether the fix is
correct.