Make data kind configurable per reader
Make the kind of data returned by a reader configurable via the
setting
  data_reading.<reader>.kind

Remove special handling for reader 'fci_l1_geoobs', i.e. the hard coded
mapping of it to kind VECTORS which was introduced in commit
"Add vector data loading to SatpyImporter"
(aa68b5f6666224ac494f10137b18a6d1f7537036).

If no kind is configured for a reader it's kind defaults to IMAGE
(except for the reader 'grib', which defaults to CONTOUR, but this
special case si removed in the next commit).