Add styling for POINTS data display
Add the new Info Enum item 'STYLE' and Presentation property 'style' as
well as functionality to deal with it analogously to
'COLORMAP'/'colormap'. This includes reading style configuration,
mapping it by name based on the layer info SHORT_NAME to layers
presenting data kind POINTS and mapping the style configuration to the
Vispy node 'Marker' styling parameters.

The style configuration follows SVG/HTML styling (see
https://www.w3.org/TR/SVG/styling.html and
https://www.w3.org/TR/SVG/painting.html). Related ideas exist in Vispy,
see 'Implement subset of SVG specification in visuals #158'
(https://github.com/vispy/vispy/issues/158).

TODO: Current Vispy (version 0.6.x) does not support to define Markers
analogous to SVG-style 'fill: none', i.e. the only way to draw the
interior of a Marker completely transparent is to define a transparent
fill color (e.g. rgba(0, 0, 0, 0)). But fragments are not discared but
their Z-buffer entry is touched. The result is, that only the first
drawn Marker correctly lets the background shine through while
overlapping markes drawn later don't render into the interior of the
previous one. The effect is, that the first drawn marker seems to be in
front of the later drawn one hiding parts of it.
The implementation contains TODO comments to mark, where it needs to be
adjusted once Vispy supports Markers with unfilled interior (the same
holds for the stroke/outline of the Marker).