Make determining POINTS layer name more robust
In contrast to the IMAGE layers the layer info of POINTS layers does not
(yet) have an entry for Info.STANDARD_NAME, thus for now a fallback
strategy is implemented which attempts to find a name in the layer info
harder.

In the future there should be an agreed strategy how to deal with
unset layer name(s) for all layer types (including IMAGE).