Fix memmap for uint16 IMAGEs
It was falsely assumed in the commit "Use memmap related function for
kind IMAGE too" (d0ee9306e09c9d7b57efe376b60a749ed2c7b20d) that the
original data type of an IMAGE could be dealt with, but data type
conversion to hard coded dtype=np.float32 is necessary for now (very
dubious!) for kind IMAGE.

Without this enforced type conversion to float32 '(pixel) counts' image
data which is of type uint16 leads to:

  ValueError: mmap length is greater than file size

in ActiveContent._attach().mm(), apparently because a memmap with dtype
float32 is expected there while it was created with only unit16. So the
actual issue seems to be, that data type handling/conversion is too
inflexible/somewhat implict here (Recommendation: revise that!).

Note: the method 'create_image_file_cache_data()' removed with the
previous commit was the result of earlier refactoring but still it makes
sense to evolve this into a commonly used function
'_create_data_memmap_file()' which deals with: creating the memmap file.