Fix switching between polygon region and full data probing
When switching from polygon to full data the region polygon was still
drawn and when drawing a new polygon while full data selection was on it
didn't work correctly either.

Thus the implementation needed revision.

Encountered problem during implementation: When attempting to just call
both ProbeGraphManager.currentPolygonChanged() and
ProbeGraphManager.currentFullDataSelectionChanged() with according
parameters to switch from one mode to the other, exceptions and broken
graphs occurred: both the methods trigger a graph rebuild in the
background thus there were two rebuilds for one graph running at the
same time and interfering.

This led to the Enhancements:
- Merge methods
    ProbeGraphDisplay.setFullDataSelection()
    ProbeGraphDisplay.setPolygon()
       to new method
    ProbeGraphDisplay.setRegion()
  to make it harder to break the invariant, that there cannot be both a
  region polygon and a full data selection per graph at the same time.
- Rename the revised currentPolygonChanged() method to
  current_graph_set_region() because it is not used as Qt slot (any
  more?) thus the "Changed" suffix is misleading.