Delegate restart of app to systemd
In the old model the MTG-SIFT application was restarted as a subprocess
of the watchdog. In order to better integrate with the monitoring
environment, we now use systemd for the restart.

In the new model MTG-SIFT and the watchdog are started as user services
by systemd. The watchdog still issues the restart request with SIGUSR1.
If the user agrees, the MTG-SIFT application will exit. Then systemd
will start again because of `restart=always`.