Handle Timebase management via TimebaseModel
Functions for finding the nearest past datetime or the index of the
nearest past datetime out of a list of reference datetimes now have type
annotations with the return type Optional, as they may return None if no
appropriate datetime can be found.
This concerns time matching policies and a function that is called when
changing the timebase from one data layer to another, as the most recent
past matching datetime in the new data layer is sought as the new
current t_sim. If no such datetime can be found the first datetime of
the new timebase's timeline is assigned to t_sim.