Introduce AreaDefinitionsManager
The AreaDefinitionsManger shall become responsible for managing all
projection configurations i.e. area definitions to be made available in
the application.

The list of area definitions offered to the user is a subset of those
provided via Satpy configuration. The MTG-SIFT configuration contains a
name for each enabled area definition and the associated area_id. The
name - you may call it 'display name' - is used in the GUI and is
usually shorter than the original 'description' of the area definition
but less abstract than the area_id.