Create test pattern image replacing actual data
When data is loaded, intercept it before it is put into the scenegraph
and overwrite it with a test pattern but without changing its shape.

This test image then allows to verify the coordinate system setup and
alignment in context of the map projections specified in the area
definiton of the data.