Suppress resampling for same area definition
Skip resampling, if the resampling target area definition is the same as
the one of the data to be loaded/resampled, as it would resample from/to
the identical grids otherwise.

The comparison is based on identity of the *area_id* of the involved
AreaDefinitions, no further attempt is done to find out, whether the
two AreaDefinitions might be semantically equivalent (same grid, same
projection, ...), if their area_ids differ.