Revise watchdog implementation a bit (2)
Use list instead of str formatting for subprocess.run

The 'subprocess.run' function also accepts a sequence of str values.
'notification_cmd' has the type 'Optional[str]' and in line 127 it
is verified, that the value can't be 'None'. The variables 'machine',
'process_name', 'severity' and 'text' are passed to 'shlex.quote()'
and must therefore also be of type str.

All elements of the command are of type str. Thus the str formatting
is not needed.