Revise the resampler selection for swath data
... in OpenFileWizard.

Not all resamplers are suitable to resample sath data thus make the
selectable resamplers adapt to the geometry definition type of the
expected data.

A problem to solve here is, that the data is not yet loaded, thus there
is no scene yet to query the geometry type from it. The Satpy readers
themselves do not provide a query interface to as beforehand which kind
of geometry the data they deliver will have. Thus the same approach is
taken as for determining the data kind (IMAGE, POINTS, VECTORS, see
commit "Make data kind configurable per reader",
20246ad9024743eb5dc5bc7e98d726d317a0a117): an configuration option
'geometry_definition' is added to the readers configuration.

Implementation details:
- the information, which resampling methods are suitable for which
  geometry definition is stored in the enhanced dictionary
  RESAMPLING_METHODS. For now this cannot be modified by configuration
  but allows for easy adjustments/adding of new resampling methods.
- the internal logic to set up the resampling combobox is
  streamlined a little in addition to the geometry definition dependet
  enhancements
- rename/type some variables/method parameters to be more consistent