Change LayerModel to allow changing the underlying list
Previously empty spaces were generated in the ComboBox if entries in
the LayerModel were deleted. This is now fixed.
The list holding the salient information for LayerModel may now be
updated via a property of LayerModel.
This is then immediately used in TimeManager, instead of clearing the
list, pushing and/or popping to/from it.