Resolve resolution/shape concept discrepancy regarding resampling
... regarding resampling. In that context the term 'resolution' was used
in the implementation to denote 'number of pixels of a grid in x/y
direction' but this concept is called 'shape' in Satpy/Pyresample where
on the other hand 'resolution' means 'size of pixel (in some unit)'.

Rename everything referring to 'number of pixels of a grid' (at least in
resampling context) to 'shape'.

This has also the effect to fix an issue: the DynamicAreaDefinition used
to set up the resampling target grid/geometry gets the correct parameter
'shape' now. This enables resampling to Plate Carree.

Also assume (and implement) that 'shape' is understood as (row, column)
tuple instead of an (x, y) tuple. This also was inconsistent. Note, that
this understanding correlates to geograpic coordinates as (latitude,
longitude) tuples.