Prevent auto restart initialization in normal mode
i.e., only when the application is configured to run in auto update mode
the initialization steps to work together with the watchdog auto-restart
feature.

This is important to be able to run the application under Windows in
normal mode since all the auto update mode features are only developed
and tested for Linux usage. In detail the watchdog initialization relies
on the existence of a signal SIGUSR1 which does not exist under Windows.