Update the dataset file table when the path updates
When the user selects or enters a directory in the QFileDialog,
then the table on the first page of the OpenFileWizard2 is updated.

  https://myprogrammingnotes.com/qfiledialog-bug.html

The ShowDirsOnly option doesn't work with the Directory FileMode,
so the user is able to select regular files, but these are filtered.
See the comment in '_update_input_directory()'.