Reuse the OpenFileWizard instance
Save the object in the variable '_wizard_dialog'.

This change is needed for the following feature:
If the user opens the OpenFileWizard again, then the import
path, filter and reader should be preserved.

Also remove the local variable 'scenes', which was only used
to set the value of the 'scenes' element in the
'importer_kwargs' dict.