Fix waste of timeline space when first data is "late" in an hour
First timeline tick used to always be chosen at beginning of that hour
the first data of the driving layer is in. By removing this constraint
a more suitable starting tick of the timeline is selected when the first
data is close to the end of an hour.

Fixes #57 "Animating MSG RSS doesnt match the timeline correctly"