Fix date display, point probe update for timeline animations
by completely bypassing the LayerSet.next_frame() method. It is a relic
from the upstream concept of grouping layers to families and animating
by toggling visibility of layers of such a family.

TODO: there are more relics left not to mention LayerSet.next_frame()
itself. They are not cleaned up with this commit to not risk stability
for now, but the whole 'frame' related implementation which is tightly
linked to the layer family animation has to be removed. See #64.

Fixes #65 "Point probe values do not update if different repeat cycle is
selected from timeline"

Fixes #66 "Data and date display is not in sync with timeline during
animations"