Fix potential runtime crash in get_data_layer_by_id
The DataLayerCollection's get_data_layer_by_id method previously
may have caused a crash if for some reason the queried index is out of
bounds for the DataLayerCollection.
This has now been fixed as well as returning None if an IndexError is
caught. Other Exceptions are not handled by get_data_layer_by_id.