Implement _create_product_from_array method in SimpleWorkspace
Previously no working version of this method existed for
SimpleWorkspace, only for CachingWorkspace which used the metadatabase
(the cache) by the nature of CachingWorkspace.  Here, a equivalent
version is implemented without a metadatabase.

The SimpleWorkspace class provides the products and contents
dictionaries instead of the database used by the CachingWorkspace. The
product and the content of generated algebraic and/or an RGB image are
now activated within SimpleWorkspace by writing them under their
respective UUIDS into these dictionaries.