Rework POINTS/LINES/VECTORS infrastructure
to support the standard data layout of Satpy readers and interpret it
according to configured Kind as:
- POINTS data:
  - Point locations (geo coordinates) are taken from the lats and lons
    arrays of the SwathDefinition of the dataset
  - if selected for loading, a field is used as attribute for colouring
    the point markes when rendering, for example: 'altitude'
- LINES data:
  - Locations (geo coordinates) of start point of each line are taken
    from the lats and lons arrays of the SwathDefinition of the dataset
    (as for POINTS data)
  - Locations (geo coordinates) of end point of each line are taken
    from two configured/selected fields, for example:
    'longitude_reference', 'latitude_reference'

Implementation details:

The datasets for POINTS and LINES data must be converted to match the
data layout the according VisPy nodes need to render it. This happens in
an early stage of the convert process. The converted datasets don't
exist on on disk. Thus they can't be found in the later import process
by the instantiated SatpyImporter, if it would create an empty scene for
its
if the scene

Select correct scene instead of merging them

'from_products' recreates the scene based on the 'paths' list. The keyword
argument 'scenes' contains the scenes, which were created in the past. The
converted dataarrays are included in these scenes.

The key from the 'scenes' dict is a tuple of all input files for the scene.
We reuse the Scene only if the `paths` list and the input files for the
Scene are the same.