Reuse Scene for Importer to rescue converted datasets
Implementation details, this is a tricky one (this description may be
wrong in details, this is overly complicated stuff):

The datasets for POINTS and LINES data must be converted to match the
data layout the according VisPy nodes need to render it. This happens in
an early stage of the loading process. But since the converted datasets
don't exist on on disk they can't be found in the later import process
by the instantiated SatpyImporter, if it would try to normally load them
again. By injecting a Scene (via kwargs['scene']) the SatpyImporter is
prevented from creating an empty scene, it reuses the given one, thus
magically the converted datasets survive.  It is important though use
the right, existing scene to do this injection, thus the correct scene
has to be re-used:

The code path following 'aImporter.from_products()' would re-create a
Scene based on a list 'paths'. But in that function the keyword argument
'scenes' contains the scenes, which were created in the past and the
converted dataarrays are included in these scenes. By finding the Scene
from this list which has a files list matching the list 'paths' and
re-using it, the converted datasets are saved across.