Refactor setting of dataset attrs in load_all_datasets
... into helper methods. This makes the code much better readable and
allows for reuse of these methods at other locations, namely the
foreseen merging of data from different datasets.

There are the following semantic changes:

Possible to explicitly set SHORT_NAME, LONG_NAME, STANDARD_NAME possible
when calling the helper _set_name_metadata() with fallback of all of
them to a 'name' given or found in the dataset attrs. When called
without actual parameters (as it is in the refactored
load_all_datasets()), the behaviour should be as before.

No other semantic changes intended.