Fix implementation of SimpleWorkspace.get_content()
In the current implementation of the method the list 'content' will
contain the element 'None' if the UUID doesn't exist. The attempt to
access an attribute 'info' of that elememt (None(!)) must fail then
raising an exception.

This commit replaces the problematic code while keeping its semantics as
close as possible to the original code in the implementation of
CachingWorkspace.get_content().

TODO: in the future the implementations may be cleaned up/completed in
both SimpleWorkspace and CachingWorkspace (multiple contents per
Product/UUID seem to have been foreseen in SSEC's original Workspace
implementation) or if CachingWorkspace is dropped.

Fix thanks to the code problem pointed out by Nicolai Kellerer.