Add Qt Model View Framework class 'LayerModel'
'LayerModel' is a subclass of 'QAbstractItemModel' and is the model of
the layer management UI. It stores information about layers in
'LayerItem' instances and triggers creation of the correct scene graph
group nodes and data nodes via signal/slot mechanisms.

Furthermore it holds an instance of a policy that determines which layer
a newly loaded dataset fits into. Currently, there is only the
ProductFamilyKeyMappingPolicy, which implements a mapping analogous to
the legacy SIFT product family key based grouping.