Make algebraics work for resampled GEOS data
When input data is resampled, the according Info must reflect the target
AreaDefinition. This is the case for ShapeDefinition based input data
(due to the previous commit) but not for AreaDefinition based data.

With wrongly set Info data, especially SHAPE information that does not
match the extents of the data grid, deriving data from it cannot work.

Fix this: Also when resampling AreaDefinition based data make sure the
resampled data has the correct SHAPE of the target area.

TODO:
It still doesn't work for e.g. SEVIRI Non-HRV with SEVIRI HRV when the
latter is resampled to the Non-HRV grid. The cause: for unknown reasons
the Info.PROJ of the HRV deviates in the last two decimal places in the
parameter "reverse flattening":
  Non-HRV Info.PROJ:  '+proj=geos ... +rf=295.488065897001 ...'
  HRV     Info.PROJ:  '+proj=geos ... +rf=295.488065897014 ...'

Since the projections of all input layers are checked, among other
things, to see if they are identical to make a composite layer 'valid',
the latter is not and therefore cannot be rendered (don't ask...).