Implement opacity control widget and delegate using it
The appearance and behavior of the default QDial widget is not suitable
for displaying and changing the visibility and opacity of a layer.
Therefore, a suitable user interface is implemented in form of a
pie-chart-like QWidget (PieDialEditor) and a QItemDelegate
(PieDialDelegate).

Opacity is specified by the angle of a disk sector:
- For example, a full disk corresponds to an opacity of 100%.
- If the central angle of the disk sector is 180 degrees, i.e. it is
  half a disk, the opacity is 50%.

Visibility is indicated by crossing out or not crossing out the disk
(sector).  A crossed-out disk (sector) means that the corresponding
layer is switched invisible.

The implementation is inspired by the following code sources:
- https://pastebin.com/yN3B3fc2
- https://stackoverflow.com/questions/12011147/how-to-create-a-3-color-gradient-dial-indicator-the-one-that-shows-green-yellow