Prevent OpenFileWizard(2) from exceeding screen width
When the OpenFileWizard(2) is opened with a pre-selection for a reader
that has a very long filter pattern, the 'Filter' combobox on the first
page becomes very wide and prevents resizing to a smaller width. This
makes the whole wizard to span even two or more screens.

Fix this by changing properties of the combobox to:
- SizeAdjustPolicy: AdjustToMinimumContentsLength
- MinimumContentsLength: 150

With this, the combobox reserves space for 150 characters, but not more.