Fix broken POINTS/VECTORS data loading
which was corrupted in commits
a81b4d19 ("Make algebraics work for resampled LEO data", 2021-10-29)
and
013c145b ("Make algebraics work for resampled GEOS data", 2021-10-30)

It was forgotten, that POINTS/VECTORS (in the future: LINES) data also
comes without an 'area' and is not resampled to something with one, thus
defining the so called SCENE information cannot succed, if it relies on
an 'area'. Actually there is no sensible setting for SCENE (in the
meaning it seems to be used) for such data, thus: let it be 'None'.

To make the implemented logic clearer and have a location to document
this in the code: refactor the logic into a method
_compute_scene_hash().