Prevent granule merging cornercase(2): into resampled
If a multi-file dataset has been loaded partially with resampling and in
a second loading turn missing files from that dataset are tried to be
loaded without resampling, these cannot be merged into the previous one.
The according test tries to detect this case by checking, whether the
SHAPEs are compatible for broadcasting the second data array into the
first one.

TODO: This condition is necessary but not sufficient. Better: add
explicit metadata e.g. to the FAMILY showing, that the data has been
resampled?!?