Replace Layer Manager QTreeView by custom LayerTreeView
For the PieDial, it is required that a PersistentEditor is open and
the editing is active all the time and does not need e.g. an
additional click to activate. So an inherited class of QTreeView is
needed and this Custom TreeView must open a PersistentEditor in the
overridden method rowsInserted.

(cherry picked from commit 8e13c3ba7a19a1da7d0f41375e863dadbaa1db94)