Avoid padding to full-disk in first load pass
Loading non-contiguous granules *without padding* (i.e. pad_data=False)
yields a Satpy Scene with a StackedAreaDefinition, that does not have an
'extent'. The 'extent' is needed after the first loading pass(*) though.

But since the StackedAreaDefinition in this case only contains
'compatible' AreaDefinitions sharing the same 'area_id' of a
pre-configured AreaDefinition, it is possible to retrieve the latter
AreaDefinition by using satpy.resample.get_area_def with a `area_id` and
use its 'extent'.

Doing this, data padding can be avoided in the first loading pass. This
reduces loading times significantly and is less memory consuming too.

Note, that some of the dataset information retrieved after the first
loading (metadata retrieval) pass without padding may not be consistent
with the dataset information of the actual loading pass. It must then be
updated. Currently indentified discrepancy (TODO: verify completeness of
this list):
   - prod.info[Info.SHAPE]

(*) It is another story whether it is clever why and how this
information is used and needed

TODO: Actually the introduced staticmethod '_get_area_extent' should be
provided as convenience function for StackedAreaDefinition by Satpy.