Fix dask TypeError in '_stack_data_arrays'
When converting POINTS, 3 DataArrays are passed into the
`_stack_data_arrays` function. The longitude and latitude DataArrays
have the data type `float32`. The `peak_current` DataArray has the
type `int16`.

In version 2.3 of dask the function `dask.array.asarray` converted a
`xarray.DataArray` to a `dask.array` independent of the data type.
However, in version 2021.11.2 an `int16` DataArray is converted into
`dask.Series`.

All DataArrays are converted to the same dtype in `dask.stack`, but
one DataArray is by now actually a Series. Thus the invariants for
the rest of the `stack` method are invalid. The exception is raised
in line 4811 in `dask/array/core.py`:

    seq2 = [a for a in seq if a.size]

    TypeError: Trying to convert dd.Scalar<size-ag..., dtype=int64>
    to a boolean value. Because Dask objects are lazily evaluated,
    they cannot be converted to a boolean value or used in boolean
    conditions like if statements. Try calling .compute() to force
    computation prior to converting to a boolean value or using in
    a conditional statement.

Solution: convert all DataArrays to float32 before calling `da.stack`