Enhance segment merging for TileGeolocatedImage display
The clean way to inform the system about a change in the data of a
dataset referenced by a scenegraph node for rendering is to emit an
according signal and connect it to a slot update_basic_layer(), which
updates the nodes data.

It's implemented that way even though for modifying the data in-place
seems not to require to call node.set_data() on the first glance.

Unfortunately for the node type, where a way to inform the node about
updated data would be necessary, node.set_data() does not work:
TileGeolocatedImage has no way to change its data (set_data() raises a
NotImplementedError exception) or to be notified, that it has to drop
all it internal tiled data.

The current commit works around this a little: it supresses the
exception for TileGeolocatedImage with fingers crossed. That works
partially: because the data is changed in-place, the node renders it -
but only for tiles created after the data modification - zooming and
panning may update some of the renderings, but when zooming and panning
you may get to see outdated tiles not having the new data.

Conclusion: don't use granule merging with tiled display, only with
pixel matrix display.