Fix problem with caching with new Content class
The CachingWorkspace had some queries where the lod attribute of
Content is used.
After the revision of class Content and the new classes inheriting
from it due to the POINTS/LINES redesign (see commit
d84e2a0c ("Add Content subclasses to metadatabase", 2021-10-11),
the loading of the data does not work anymore, because the
class Content does not have the attribute lod anymore, but the
inherited class ContentImage has this now.

A check is therefore needed to see whether a ContentImage must be
used or not so that loading works again.