Fix regression when caching is on due to merging
When caching is turned on (the document workspace is a
CachingWorkspace), sorting (*) the uuids based on a query of the
workspace's underlying database breaks the implicit correlation between
uuid[s] and merge_uuid[s], since they are independent lists and only one
of them is reordered. Even if merging is turned off, breaking this
correlation will cause the wrong code path to be processed, since the
criterion for performing a merge is the result of the "uuid !=
merge_uuid" test (reordering causes this test to be negative at most by
chance).

Therefore:
- Replace the merge_uuid[s] list with a dict that maps uuids to the uuid
  of the corresponding merge targets to make the correlation explicit.
This allows the individual uuids of the list to be reordered, and the
uuid of a corresponding merge target can still be determined.
- Secure the code path that does the merging so that it can only be
  reached when the merging is active.

(*) it is actually completely unclear to me what this sorting is
supposed to be good for. Presumably it will finally become obsolete as
soon as the new layer management design is introduced.