Add Recipe 'id' partly taking over the role of 'name'
Recipe.id shall be used to identify the recipe internally when
referencing it in dependency calculations.

The Recipe.name can eventually be used to identify the recipe in
displays or when saving it to disk (future).

Implementation details:
The identifier 'Recipe.id' shall not be part of the constructor for two
reasons:
1. keep interface compatibility to the namedtuple based initial
   CompositeRecipe implementation. Just to reduce code changes at other
   locations.
2. since the id is meant to be an internal detail of the class it should
   not be set from outside.

TODO: there are two disadvantages with the current implementation:
- Recipe.id mutable from outside the class
- Recipe.id is written to the dict via Recipe.to_dict(), which doesn't
  make sense